% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_test_cop.R
\name{cond_cov}
\alias{cond_cov}
\title{cond_cov}
\usage{
cond_cov(Met, nv)
}
\arguments{
\item{Met}{A numeric matrix, usually a covariance matrix}

\item{nv}{An integer indicating the location of what is to be inverted.}
}
\description{
This function is used to compute the conditional covariance when using the
\code{diagMAAR.cop} method.
}
