% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_test_cop.R
\name{extract_stat}
\alias{extract_stat}
\title{extract_stat}
\usage{
extract_stat(Vec, alpha = 0.05, med = FALSE)
}
\arguments{
\item{Vec}{A numeric matrix, usually a covariance matrix}

\item{alpha}{An numeric value specifying the upper and lower quantiles}

\item{med}{A logical value, if TRUE the function returns the medium instead 
of the  the mean.}
}
\description{
This function is used to extract the mean, medium, upper and lower quantiles
(as specified by alpha) of the input vector.
}
