% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_prob_function.R
\name{logistic_prob}
\alias{logistic_prob}
\title{Logistic Probability}
\usage{
logistic_prob(beta0, Xbeta, mu = NULL)
}
\arguments{
\item{beta0:}{A numeric value.}

\item{Xbeta:}{A numeric matrix.}

\item{mu:}{A numeric value, by default it is set to NULL}
}
\value{
p: logit^{-1}(beta0 + Xbeta) if is.null(mu) or 
mean(logit^{-1}(beta0 + Xbeta)) - mu.
}
\description{
This function computes \eqn{logit^{-1}(\beta_0 + X\beta)} if is.null(mu), 
otherwise it computes \eqn{mean(logit^{-1}(\beta_0 + X\beta) - \mu)}.
}
\details{
This function is used when generating missing data from a complete data set.
}
\seealso{
Other data generating functions: \code{\link{MAAR_mechanism_noindep}},
  \code{\link{MAAR_mechanism}},
  \code{\link{MANAR_mechanism}}, \code{\link{sample_mvn}}
}
