% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_function.R
\name{prep.dtmm}
\alias{prep.dtmm}
\title{Prepare for directly testing a postulated missingness mechanism}
\usage{
prep.dtmm(dt)
}
\arguments{
\item{dt}{A numeric matrix, data frame or data table.}
}
\value{
prep A prep S3  object that contains the original data,
the desired daigtest to use, the sample size, the number of missing
variables, the location of the missing variables, the number of variables and
a data set that has the missingness indicators appended.
}
\description{
\code{prep.dtmm} takes a data frame, or matrix and prepares it to be analyzed
 by diagMAAR.dtmm.
}
\details{
\code{prep.dtmm} is part of the data preparation functions used to pre-process
data sets so that diagnostic tests for the MAAR assumption can be ran.
}
\examples{
# Generate 100 iid samples from a MVN with correlation equal to 0.3
samples.mvn <- sample_mvn(5, 0.3, 100)
# Take the Gaussian data and and delete some values from the fourth row.
obs.nvmn <- MAAR_mechanism(samples = samples.mvn, miss.coef = 0.2, 
                           miss.nvar = 1, miss.var = 4, 
                           prob.coef = matrix(c(-1, 0.5, 0.7, - 0.2), 1, 4))
dt.prep <- prep.dtmm(obs.mvn)
print(dt.prep)
}
\seealso{
Other prep: \code{\link{prep.ccm}}, \code{\link{prep.cop}},
  \code{\link{prep}}
}
