% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mvn_function.R
\name{sample_mvn}
\alias{sample_mvn}
\title{Sample a Multivariate Normal}
\usage{
sample_mvn(nvar, rho, N, seed = 0)
}
\arguments{
\item{nvar}{An integer specifying the dimension of the multivariate normal
distribution.}

\item{rho}{The correlation matrix, if a single number is given then the
correlation is assumed to be constant between all of the variable.}

\item{N}{An integer specifying the number of samples.}

\item{seed}{The seed used to generate that data set.seed.}
}
\value{
samples.mvn A matrix of dimensions equal to (nvar, N) containing the
        samples from a MVN.
}
\description{
This function creates N samples from a multivariate normal distribution
with mean equal to zero and pairwise-correlation rho.
}
\details{
This is a basic functions that samples a standard normal and uses the
Choleski decomposition to create correlated samples.
}
\examples{
sample_mvn(5, 0.3, 100)
}
\seealso{
Other data generating functions: \code{\link{MAAR_mechanism_noindep}},
  \code{\link{MAAR_mechanism}},
  \code{\link{MANAR_mechanism}},
  \code{\link{logistic_prob}}
}
\keyword{Multivariate}
\keyword{Normal}
\keyword{Sample}
